sha1_update() shouldn't overwrite caller buffers
The sha1_update() function, when given a buffer of more than 64
characters, would copy the first 64 characters into its internal buffer,
and the first s_transform() would be done on that.  Subsequent
transforms were done in-place, on the user-supplied buffer.

This caused the buffer data to become corrupt.  The resulting checksum
value was correct, however, which is why it wasn't initially detected by
the test suite.

A new tests has been put in place to verify that sha1_update(), via its
more common public interface sha1_data(), is not corrupting arg buffers.

v1.2.5